Add data-type="text" to points_lines_and_planes for mobile testing.
Summary:
The iOS app isn't able to detect that the text box in
points_lines_and_planes is supposed to allow letters and not just
numbers, which hopefully this should help. We can't test that this works
for the mobile app because there's a bug with local testing, but this
does work on webapp and shouldn't affect mobile either.

Test Plan:
 - Visit
   http://exercises.ka.local/exercises/points_lines_and_planes.html?debug&problem=name-plane
 - See that the question is still answerable.

Reviewers: bryan, eater

Reviewed By: eater

Differential Revision: https://phabricator.khanacademy.org/D20501