Remove ServerActionQueue from interface.js.
Summary:
The ServerActionQueue was added for the [hint cheating project](https://docs.google.com/a/khanacademy.org/document/d/1gLXa3kD89mZtcms4p0SYtDrY3lgRY_nfPrNuE6SZP6E/edit?usp=sharing)
but it unfortunately caused some nasty bugs in production. After trying
to clean things up this week I've come to the conclusion that the queue
(in any form) is very dangerous and shouldn't be done right now.

This change removes the queue and reverts all changes related to it.

In a follow up diff I will solve the original objective of the hint
cheating project by safer/simpler means.

Test Plan:
This is removing code from paths that were not being executed because
no one was allowed to pass the gandalf bridge. So this should be a very
safe change.

Run through an exercise on your mission dashboard and ensure that
everything behaves. Be sure to take some hints.

Reviewers: alex, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D20934