Remove MathZoom.js from MathJax
Summary:
In a previous diff I added MathZoom.js from MathJax because some browsers were
requesting.  This diff removes the file and removes it from the list of extensions
we might load.  In a follow up diff I will update references to the old MathJax
config in webapp so that they're pointing at this new build.

Test Plan:
- load http://localhost:8080/math/algebra/systems-of-linear-equations/introduction-to-systems-of-linear-equations/e/verifying-solutions-to-systems-of-equations?mp-r-id=yqKY-Ao=
- verify that MathJax.js loads but that the browser doesn't load (or try to load) MathZoom.js

Reviewers: emily

Subscribers: eater

Differential Revision: https://phabricator.khanacademy.org/D21775