Fix rounding whole numbers
Summary:
Users are stuck on rounding whole numbers because the place they are supposed to round to is not being displayed.

Perhaps the problem is with casting TPLACE to a string. This does that manually.

The relevant asana task is

https://app.asana.com/0/27216215224639/49975697797334

Before this diff, debug information would display
TPLACE as {}, now it will display as "ten" or whatever the appropriate number is. TPLACEPLURAL will be "tens" or the corresponding number.

I hope this works - I wasn't able to repro the bug on my machine, and no one at KA has been able to. I'll follow up a day after deploy and see if this has fixed the problem. I suspect it will.

Test Plan:
The debug information seems more sensible now - instead of logging TPLACE as {}, it logs it as the appropriate number.

I wish I could repro, but so far no one's been able to do so. I think it makes sense to deploy this and see if the problems go away. Another option would be to replace this exercise with a perseus exercise with the same content.

Reviewers: eater, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D21863