Temporarily don't detect offline hints.
Summary:
Hopefully fixes https://app.asana.com/0/27216215224639/49826201680046

The plan is to monitor Zendesk for a day after this is deployed.
If this reduces the number of reports of hints being seen
substantially, we con probably fix this by including the item
sha in the hint detection and we can do this. A real repro has been
a bear to reproduce, so this test plan is the best we can do.

Test Plan:
In Firefox, Chrome, Safari, and incognito for each of those:

1. Open an exercise on your learning dashboard that has hints.
2. Kill your dev server
3. Take a hint
4. Refresh the page to show a blank page
5. Start your dev server
6. Refresh the page to show the exercise again
7. Verify there are no errors (though stil no hints)
8. Congrats! You're a cheater

Reviewers: johnsullivan

Reviewed By: johnsullivan

Differential Revision: https://phabricator.khanacademy.org/D22128