Add a bunch of eslint-disable comments for legacy code.
Summary:
We're introducing a new, stricter js/jsx linter in the khan-linter
repo: eslint.  It has a bunch of rules to enforce our style guide,
that the old jshint did not have.  But most legacy code doesn't follow
our style guide so well.

This change allows us to upgrade to the new linter for new code, while
keeping old code compliant.  I do this by adding 'eslint-disable'
directives everywhere.

Here's the mondo command I ran to do this all automatically.  First, I
applied D22374 in my khan/devtools/khan-linter rule.  Then, in
khan-exercises/, I ran this mondo command:

  ~/khan/devtools/khan-linter/runlint.py > /var/tmp/le
  cat /var/tmp/le | grep -v 'File ignored' | perl -nle '/^([^:]*):\d+:\d+: [EW]([^ ]*)/ and print "$1 $2"' | sort -u | perl -ane '$t{$F[0]} .= " $F[1],"; if (eof) {for $k (keys %t) { chop $t{$k}; open(OLD, $k); open(NEW, ">$k.tmp"); print NEW qq{/* TODO(csilvers): fix these lint errors (http://eslint.org/docs/rules): */\n/* eslint-disable$t{$k} */\n/* To fix, remove an entry above, run ka-lint, and fix errors. */\n\n}; while (<OLD>) {print NEW $_}; close(OLD); close(NEW); rename("$k.tmp", "$k") }}'

It's not as bad as it looks.  Basically it runs the linter, munges the
output to be 'filename violated-rule-name' pairs, and then uses some
perl to collect all the violated-rules for a single filename in one
place, and to prepend the appropriate eslint-disable comment on to the
file.

Test Plan:
In khan-exercises/, ran
   ~/khan/devtools/khan-linter/runlint.py
and saw no errors.

Reviewers: emily, kevinb

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D22393