Reintroduce offline cheating detection.
Summary:
We removed the offline cheating detection because it was misfiring too
often. I've come up with two potential causes of this misfiring and
fixed both of them:

1) We were not keying the `OfflineHintRecord` on the user ID (doh).
2) We were not checking the the `userExercise` object had a name (when
   content creators look at a preview in Perseus One the exercise
   name will be undefined).

Because the symptoms of this bug are somewhat severe (students being
unable to get X in a row succsefully) I'm rolling this out tentatively.
Every time we detect cheating, we'll send a message to the server that
will make its way into our logs. I'll look at the frequency of these
logs, as well as the percentage of coached users (most likely to cheat)
that are caught with this, to see if the code is working or not.

Test Plan:
1) Open up any exercise.
2) Disconnect from the internet or kill your app server.
3) Take a hint.
4) Close the tab (or refresh to get a blank screen)
5) Reconnect or bring back up your app server
6) See no hints, but see that the added Ajax request fired succesfully.

Reviewers: eliana

Reviewed By: eliana

Subscribers: #sat

Differential Revision: https://phabricator.khanacademy.org/D22672