Add support for two new problem-error types.
Summary:
In a separate commit (in webapp), we are going to change the options
that we show when you click on 'report a problem' at the bottom-right
of pages like
   https://es.khanacademy.org/math/trigonometry/less-basic-trigonometry/law-sines-cosines/e/law_of_sines

This handles the backend (khan-exercises) support for this: Mapping
the new options (indexed by element-id) to the jira name we use for
them, and to the placeholder text we associate with them.

Test Plan:
Fingers crossed -- I don't know how to test this until the webapp
changes are in too, and everything is deployed!

Reviewers: eater

Reviewed By: eater

Subscribers: cameron

Differential Revision: https://phabricator.khanacademy.org/D23140