Normalize message for unanswered custom problems.
When a custom problem is not answered, the validator is expected to
return an empty string result. The "score" is then marked as "empty".
Instead of passing the empty string as "message" further into the
framework it should be better to convert it to null right away.
This seems cleaner than handling it in the multiple-validator, and
makes custom exercises also display "more parts to answer" when a
user just clicks "Check Answer" without doing anything.