Don't point people to GitHub to report content issues
Summary:
When users report an issue with an exercise, that issue report generally gets
automatically sent to JIRA where the content team goes through and resolves
any legitimate content problems. Previously if communication with JIRA failed,
we'd ask users to submit an issue report to GitHub instead. Unfortunately as
the content team has grown and evolved, they're relying soley on JIRA and not
looking at GitHub anymore. Consequently, we shouldn't ask users to file issues
there as it will simply be a waste of their time. Instead if JIRA fails, we
should simply thank them for their feedback.

Of course if JIRA fails, this means their feedback has gone into a black hole
never to be seen again :(. For now, we're okay with that because any legitimate
problem with an exercise will elicit scores (or even hundreds) of similar issue
reports, so we don't expect this'll mean we miss out on anything.

Test Plan:
Modify `issue_reporting.py` to always return an error:
```
 def do_post(handler, url, data, headers):
+    handler.response.set_status(500)
+    return
```

Loaded an exercise page (e.g., `/e/addition_1`) and reported an error. Upon submitting, saw the helpful "Thank you for your feedback!" message and saw a 500 error in the network panel for the /jirapost POST.

Reviewers: johnsullivan

Reviewed By: johnsullivan

Differential Revision: https://phabricator.khanacademy.org/D23543