Refactored calculator UI code to a separate module
Summary:
https://app.asana.com/0/75463145497297/76446024361663

This changeset moves the code to initialize the calculator into its own module.
We do this so that we can access this code separate from all of khan-exercises.
Specifically, we'll be rendering calculators on the fly in the new bibliotron
exercise pages, and need to wire up the events accordingly.

The implementation of this code in webapp currently resembles the following,
but it is not yet up for review:

```
const React = require("react");
const calculatorTemplate = require("../../exercises-package/calculator.handlebars");

require("../../../khan-exercises/genfiles/calculator.js");
require("../../../khan-exercises/utils/init-calculator.js");

module.exports = React.createClass({
    componentDidMount: function() {
        Calculator.init();
    },

    render: function() {
        return <div
            dangerouslySetInnerHTML={{__html: calculatorTemplate()}}
        />;
    },
});
```

Specifically, we require two files from khan-exercises, which create a
`Calculator` global and attach an `init` function to it. We can call this
`init` from `componentDidMount`.

Test Plan:
* python -m SimpleHTTPServer
* http://localhost:8080/exercises/standard_deviation.html
* Verify the calculator works

Reviewers: michelle, emily, benjaminpollack

Reviewed By: benjaminpollack

Subscribers: jlfwong, #bibliotron

Differential Revision: https://phabricator.khanacademy.org/D24396