Make sure a string is passed into JSON.parseString
Summary:
In D23651 we added a try/catch to remedy a strange localStorage bug. In
doing so we accidentally removed some code to make sure we were passing a
string to JSON.parseString.

This bit of code adds it back, and also contains the output of `make genfiles/calculator.js`.

Test Plan:
* python -m SimpleHTTPServer
* http://localhost:8000/exercises/standard_deviation.html
* Verify the calculator works

Reviewers: benjaminpollack

Reviewed By: benjaminpollack

Subscribers: eater

Differential Revision: https://phabricator.khanacademy.org/D24433