Prevent race condition in calculator modules
Summary:
Our screenshot cron job is failing because of page errors like: http://sandcastle.kasandbox.org/media/castles/Khan:master/exercises/addition_1.html?debug&seed=31

The problem is that `Calculator` is not defined when `init-calculator.js` is
invoked. This is because we need `genfiles/calculator.js` to first be loaded,
but in D24396 I unfortunately listed the dependencies in a single array:

    require(["./genfiles/calculator.js", "./utils/init-calculator.js"], function() {
        Calculator.init();
    });

Which does not guarantee that they are loaded in order. Instead, I now nest them.

    require(["./genfiles/calculator.js"], function() {
        require(["./utils/init-calculator.js"], function() {
            Calculator.init();
        });
    });

There is a better way to do this with requirejs configs but I do not believe it
is necessary.

Test Plan:
I inserted a breakpoint just before this line on `khan-exercise.js` in http://sandcastle.kasandbox.org/media/castles/Khan:master/exercises/addition_1.html?debug&seed=31, then modified the code to use the nested require statements.
Doing so reliably prevented the console errors.

Reviewers: benjaminpollack, eater

Reviewed By: eater

Subscribers: #bibliotron

Differential Revision: https://phabricator.khanacademy.org/D24511