Properly clean up after a problem when using interfaceFunctions
Not having this line was causing a bug in webapp where a new
ItemRenderer was being added to the page every time you advanced
to the next question, and each one wanted to try and control the
same DOM nodes (this didn't work out very well).

Test Plan:
In webapp, skip questions while keeping an eye on the React dev tools.
Note that additional ItemRenderer instances are no longer being added.

Auditors: michelle