Fix angle_types.html to use a global 'angle' object instead of a DOM node
Summary:
angle_types.html was relying on the fact that `window` creates properties
based on DOM node ids.  In this case just having a DOM node with the id "angle"
set `window.angle` to point to that DOM node.  This doesn't work when using
danae.js to scrap the page because it never creates a DOM node with that id.
To fix this we instead create a global `angle` object on `window`.

Test Plan:
- open angle_types.html in the browser
- check that `window.angle` is an POJO and not a DOM node

Reviewers: emily, sergei

Reviewed By: sergei

Differential Revision: https://phabricator.khanacademy.org/D25196