Update exercises with Adders, Subtractors, Multipliers, Dividers to work with danae
Summary:
We need to make sure all of the hints are shown.  Since the Adder, etc. is created
within graphie code which isn't run because danae use the &nographie option we need
to do something different.  This diff initializes DUMMY (or HINTS1) within the "vars"
block to be an array with length equal to the number of hints that the Adder, etc.
will generate.

Test Plan: - generate questions for each of the modified assessment items and verify them in perseus (work in progress)

Reviewers: sergei, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D25427