Fix timestamp on attempt data to work across locales.
Summary:
See the inline comment.
Fix for: https://app.asana.com/0/27216215224639/92638698970527

(I added a few people who committed to khan-exercises recently, let me know if there's someone else who should be reviewing this change!)

Test Plan:
In three versions of an exercise:
http://localhost:8080/math/early-math/cc-early-math-counting-topic/cc-early-math-counting/e/counting-out-1-20-objects
http://localhost:8080/math/early-math/cc-early-math-counting-topic/cc-early-math-counting/e/counting-out-1-20-objects?lang=es
http://localhost:8080/math/early-math/cc-early-math-counting-topic/cc-early-math-counting/e/counting-out-1-20-objects?lang=bn

I dragged an apple into the bin and clicked the submit button. With the console open I confirmed that the HTTP request to the attempt API was successful. Inspecting the HTTP request I confirmed that it included: `client_dt: 2016-02-22T16:46:37-05:00`, which is what we want across all locales.

Reviewers: alex, johnsullivan, emily

Reviewed By: emily

Differential Revision: https://phabricator.khanacademy.org/D25503