Disable autoproceed behavior when registering a skip attempt
Summary:
On mobile, when the user hits the skip button, we do some work ourselves to proceed to the next
problem. (For example, we may need to load the next Perseus exercise, or tear things down and build
them back up if we're switching exercise frameworks.) The automatic trigger of `gotoNextProblem` in
`handleAttempt` is causing all sorts of bugs, so we'd like to disable it.

Instead, clients are now expected to trigger the loading of the next problem themselves, out-of-band
from the `handleAttempt` method.

Test Plan:
- Substate this change into webapp.
- Enable Bibliotron (by navigating to `localhost:8080/enable-bibliotron`).
- Navigate to a skill check (e.g., `http://localhost:8080/science/physics/one-dimensional-motion/skill-check-for-one-dimensional-motion/e/one-dimensional-motion-skill-check`).
    - Verify that skipping works!
    - Verify that checking the answer works.
- Navigate to a mission warm-up.
    - Verify that skipping works!
- For the mobile test plan, see: D25792.

Reviewers: benkomalo, jared

Reviewed By: benkomalo, jared

Subscribers: eliana, oliver, nacho, #ios

Differential Revision: https://phabricator.khanacademy.org/D25790