Update `local-only` dependencies
Summary:
D25503 included a change to use part of the Moment API that wasn't available in the version of
`moment.js` included in the `local-only` subdirectory. On mobile, these `local-only` dependencies
are actually used as production dependencies, since we don't have webapp to rely on. (This is not a
good setup, but it's what we have to work with for now.) This diff is just a result of running the
`update_local.sh` script.

Test Plan:
- On web:
    - Run khan-exercises in local mode.
    - Open up the console.
    - Verify that `moment().locale` is defined. (This behavior isn't actually used when testing locally.)
- On iOS:
    - Substate this change into the webview repo.
    - Re-build the JS bundle.
    - Verify that exercises load and complete without error.

Reviewers: jared, john

Reviewed By: jared, john

Subscribers: john, #ios

Differential Revision: https://phabricator.khanacademy.org/D25789