Require explicit ask to hide problem UI on network errors
Summary:
When we timeout on a network request to record a problem attempt (or hit any other error during
that request), khan-exercises hides the entire problem UI by applying a CSS change.

This makes for an awkward experience on mobile, where we throw up an error modal and kick the user
out of the exercise, and the exercise in the background goes blank at the same time. It can also
lead to incompletable exercises if, for whatever reason, we fail to recycle the webview on mobile,
since whenever the user opens a subsequent exercise, the UI will be blank. (There are cases where
this happens--they're being addressed independently.)

This diff tries to remove more implicit webby behavior from `interface.js` by adding a custom
error-handling callback to `handleAttempt`. Web clients will continue to see the same behavior,
while the API that mobile clients rely on will be able to pass in a `null` callback to avoid
messing with the problem UI.

Test Plan:
- Change the request timeout to something ludicrously small (1000ms is good enough).
- Substate this change into the iOS app and rebuild the JS.
- Run the mobile app.
- Navigate to an exercise.
- Attempt a problem.
    - Verify that the error modal appears, and that the exercise does _not_ blank out in the background.
- Substate this change into webapp.
- Attempt a problem.
    - Verify that the UI goes blank, etc.

Reviewers: jared, benkomalo

Reviewed By: jared, benkomalo

Subscribers: #ios

Differential Revision: https://phabricator.khanacademy.org/D25875