Rewrite adding_fractions_with_common_denominators to use \overbrace
Summary:
The reason why I went with \overbrace is that it's easier to let
MathJax handle the math than try and figure out to differences in
font size and layout between the rendering on graphie-to-png and
the rendering of the math within perseus.

Test Plan:
Load adding_fractions_with_common_denominators.html a couple of
times and see that the hints use \overbrace.

Auditors: sergei emily