Be more conservative about saying something is a multiple of pi.
Summary:
I can't figure out how 'pi' is being set as a possible answer-form for
questions like
   https://www.khanacademy.org/math/cc-fourth-grade-math/cc-4th-place-value-rounding/cc-4th-rounding/e/rounding-whole-numbers-2

But it is, and rather than have to fix that up, and then trust editors
to properly say "this answer involves pi" at all, let's just be better
about guessing whether a number is a multiple of pi.

This is probably not perfect, but I think is good enough.  If we're
worried this will suppress legitimate messaging, we could just
suppress it if the answer is an integer, which I expect will catch 95%
of the confusing cases now.

This resolves https://app.asana.com/0/27216215224639/96623486099021/f

This also resolves T2578

Test Plan:
Ran `python -m SimpleHTTPServer` and visited localhost:8000/test

Also will try the rounding-whole-numbers questions after this is deployed.
I expect not to get a "approximated pi" warning if I say xxxx0001 as
an answer.

Reviewers: eater

Reviewed By: eater

Subscribers: jared

Maniphest Tasks: T2578

Differential Revision: https://phabricator.khanacademy.org/D26043