Deal with the fact that getUserProfile isn't defined on KA anymore.
This finishes the task done in D25801.

This is not an ideal solution, since it means that powerUser and vip
will never be set anymore, but I'm not going to worry about it that
much since I think all this code is going away anyway when we move to
perseus.  (Though maybe not?  I don't know how perseus does user
feedback.)

This fixes https://app.asana.com/0/27216215224639/103422119655835/f

Auditors: emily, jlfwong

Test plan:
Fingers crossed