Parameterize HTTP requests on a retry strategy
Summary:
In our native apps, where connectivity is often inferior to that exhibited by a desktop, we want to
make exercise-related network requests automatically retry a few times on failure (and then retry
again if the user chooses to retry, rather than get kicked out of the current exercise). As such,
we need a way to configure the retry behavior for network requests kicked off in khan-exercises.

This diff adds a `retryStrategy` that is threaded down to our network requests. The `retryStrategy`
is formulated as a function that takes an index (the number of times the request has been retried)
and returns a promise that resolves to `true` if the request should be retried. This approach
allows for asynchrony (i.e., such that we can throw up a modal to the user asking them if they want
to retry) and meshes well with the current promise-based approach for network requests.

As the default strategy is to never retry, behavior should remain unchanged outside of the native
apps.

On Asana @ https://app.asana.com/0/107925931410679/107969342010003.

Test Plan:
- (Testing for the retry behavior will follow shortly in a separate diff via the Electron bridge.)
- To guarantee backwards compatibility:
    - Substate this change into webapp.
    - Navigate to an exercise.
    - Check answers, take hints, etc.
        - Verify that the requests go through successfully via DevTools.
    - Modify `request` such that the URL argument hits something live on the web (any HTTP URL), rather than your local devserver.
    - Turn off Wi-Fi.
    - Reload the page; take a hint.
        - Verify that failed requests appear in DevTools.
    - Check your answer.
        - Verify that the red bar appears (i.e., the request fully failed).

Reviewers: jared, benkomalo

Reviewed By: jared, benkomalo

Subscribers: #android, #ios

Differential Revision: https://phabricator.khanacademy.org/D26372