Remove a bunch of unused code from khan-exercises!
Summary:
This is due to the fact that khan-exercises-style exercises are gone
from the ka website.  We only need to keep the khan-exercises code
that is still used by perseus.  I could remove

1) Triggers that are not called anywhere in webapp anymore, now that
   the perseus version of the trigger is called instead.
2) All code behind localOnly.

The first was the big win, since lots of code was only being called
from those triggers.

Test Plan:
In webapp (with these changes in the khan-exercises/ subdir) ran
`make serve` and then visited
   http://localhost:8080/math/geometry/intro_euclid/e/drawing-lines
It rendered properly.  I clicked on 'show hint' and it rendered
properly too.  I then solved the problem and was able to go on to the
next one.

I then clicked on 'report a mistake in this problem' and a popup
popped up.  I clicked on 'show answer' and it showed all the hints.
So it looks like that part is properly working as well.

Reviewers: emily

Reviewed By: emily

Subscribers: charlie, jared

Differential Revision: https://phabricator.khanacademy.org/D27209