Update `local-only` version of KAS
Summary:
The mobile apps pull in KAS through the `local-only` copy bundled with khan-exercises. In D27242, a
method was added to KAS that the apps will soon need; as such, I'm updating the bundled copy so
that the mobile apps will have access as soon as we substate.

Buckle your seatbelts, because this `local-only` version of KAS hasn't been updated since November
2014! (Aside: I'm surprised we haven't seen more breakages as a result of this staleness; that
means the mobile apps have been missing ~30 commits for a few years now.)

Test Plan: Coming in the mobile substate revision.

Auditors: alex