Update color palette to match Perseus
Summary:
In D27731, Perseus's color palette was updated. Though Perseus no longer relies on khan-exercises,
graphie-to-png does, and in particular, it uses the color palette defined on `KhanUtil` when
creating images. It would be a large undertaking to remove graphie-to-png's dependency on
khan-exercises (see this [Asana ticket](https://app.asana.com/0/27216215224639/36842953088193)). As such, it seems wise to update
the color palette here (and cherry-pick the change into graphie-to-png) in the short-term.

Test Plan: Coming in the graphie-to-png integration diff (since `local-mode` doesn't seem to work anymore).

Reviewers: benkomalo, david

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27732